feat: Two types of axis components, with specialisations
The default `axis` is backwards compatible (works for all scales), and has an additional `tickCenterLabel` property for positioning the label between ticks. This can be useful for continuous scales, but is less useful for ordinal scales (example included)
