+++ packages/d3fc-axis/examples/center.html
@@ -1,0 +1,18 @@
+<html>
+<head>
+    <script src="../node_modules/babel-polyfill/dist/polyfill.js"></script>
+    <script src="../node_modules/d3/build/d3.js"></script>
+    <script src="../../d3fc-data-join/build/d3fc-data-join.js"></script>
+    <script src="../../d3fc-rebind/build/d3fc-rebind.js"></script>
+    <script src="../build/d3fc-axis.js"></script>
+    <style>
+        svg {
+            display: block;
+            margin: 10px;
+        }
+    </style>
+</head>
+<body>
+    <script src="center.js"></script>
+</body>
+</html>
