+++ packages/d3fc-axis/examples/offset.js
@@ -5,6 +5,14 @@ var scale = d3.scaleBand()
   .domain(['Carrots', 'Bananas', 'Sausages', 'Pickles', 'Aubergines', 'Artichokes', 'Spinach', 'Cucumber'])
   .range([margin, width - margin]);
 
+var linear = d3.scaleTime()
+  .domain([new Date('2019-03-02'), new Date('2019-03-07')])
+  .range([margin, width - margin]);
+
+var bands = d3.scaleBand()
+  .domain(['Carrots', 'Bananas', 'Sausages', 'Pickles'])
+  .range([margin, width - margin]);
+
 var axis = fc.axisBottom(scale)
   .decorate(function(s) {
       s.enter().select('text')
@@ -13,9 +21,25 @@ var axis = fc.axisBottom(scale)
         });
   });
 
+var axisLinear = fc.axisBottom(linear)
+    .tickArguments([5])
+    .tickCenterLabel(true)
+    .tickPadding(5)
+    .tickSizeInner(10);
+
+var axisBetweenTicks = fc.axisOrdinalBottom(bands)
+    .tickPadding(5)
+    .tickSizeInner(10);
+
 var svg = d3.select('body').append('svg')
     .attr('width', width)
-    .attr('height', 80);
+    .attr('height', 200);
 svg.append('g')
     .attr('transform', 'translate(0, 10)')
     .call(axis);
+svg.append('g')
+    .attr('transform', 'translate(0, 80)')
+    .call(axisLinear);
+svg.append('g')
+    .attr('transform', 'translate(0, 140)')
+    .call(axisBetweenTicks);
