+++ packages/d3fc-axis/index.js
@@ -1,1 +1,2 @@
 export * from './src/axis';
+export * from './src/axisOrdinal';
